requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0,<6.0
django-cors-headers>=3.7,<4.0